Hook up PinSetter results to ActionMaster and the animator.
Count the number of pins knocked down and send that to the
ActionMaster to determine the right action. Pass that action
to the animator.

Note that the pins knocked down is calculated using the last
number of settled pins, which is also reset every time we spawn a new
set of 10 pins.