Handle basic strikes and spares.
If we have a strike or spare, see if we have enough future bowls to calculate the score,
and add if so. If we do not have future bowls, then don't add a new frame score.

We shouldn't be able to accidently add the wrong frames, so there is no need
to insert a placeholder in the list.