Format bowls array as a string of chars to display.
The bowls array contains the pin bowls numbers, so we format
them to standard bowling chars (/, X, -) in a string with each char
corresponding to one bowl display box. Spaces are used when there is a strike,
since the bowls array does not contain "empty" bowls due to strikes (except when
the strike occurs in the last frame).