Add view zoom, helicopter, keybinds, and sounds
View zoom is triggered on zoom hotkey and changes fov to 1.5x smaller.

Add Helicopter game object, which maintains whether or not it
has already been called.

Helicopter can be called with H hotkey, which also plays a sound
on the player's radio.

The helicopter also has a rotor sound attached.