Add a day night cycle.
The direction sun object is now being rotated each frame. We
calculate the number of angles it should move per second (360 / (second in
a day). Then multiply a scale to make it faster.

Each frame, it will rotate an amount determined by the time since the last
frame (which is also the  time since the last rotation).