Add helicopter descent.
If the helicopter has reached the above landing zone point, move towards the real landing zone.

Logically, once the helicopter starts descent, it is no longer equal to the above point and the first
block should be true,but in game this doesn't happen and the heli descends correctly. Something to
keep an eye on though as a possible bug.