Fix day night lighting
Add reflection probe and probe light group so that things are lit correctly when the sun is tranisitioning from day to night.

We start the sun at the night phase since otherwise it seems that far away objects are light
like daytime, due to some range issue on the reflection probe that I haven't figured out.